Use React effect hooks to fix up article section headings.
Kuma runs two JavaScript modules on page load to add styles and
anchors to certain section heading within the main article content.
This PR adds those features to the React version of the site. It
moves the Article component from page.jsx into its own article.jsx
file, and then adds two useEffect() hooks to run code over the
article HTML content each time the article is re-rendered.
