+++ kuma/static/js/highlight.js
@@ -1,3 +1,5 @@
+// The React version of this code is in the highlightSections() function
+// of kuma/javascript/src/article.jsx
 (function($) {
     'use strict';
 
@@ -5,12 +7,13 @@
     function highlight(targets) {
         var $targets = $(targets);
         $targets.each(function() {
-            $(this).addClass('highlight-spanned').wrapInner('<span class="highlight-span"></span>');
+            $(this)
+                .addClass('highlight-spanned')
+                .wrapInner('<span class="highlight-span"></span>');
         });
     }
 
     var $articleSubHeads;
     $articleSubHeads = $('#wikiArticle h3, #wikiArticle h5');
     highlight($articleSubHeads);
-
 })(jQuery);
